requirements.txt
bcrypt==4.0.1
boto3==1.26.53
botocore==1.29.53
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cssselect==1.2.0
cssutils==2.6.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.2
premailer==3.10.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.46
urllib3==1.26.14
Werkzeug==2.2.2
yagmail==0.15.293
zipp==3.11.0
